+++ src/main/java/io/nagurea/smsupsdk/campaigns/get/replies/Reply.java
@@ -1,0 +1,19 @@
+package io.nagurea.smsupsdk.campaigns.get.replies;
+
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+@EqualsAndHashCode
+@ToString
+public class Reply {
+    private final String destination;
+    private final String info1;
+    private final String info2;
+    private final LocalDateTime date;
+}